requirements.txt
beautifulsoup4==4.4.1
lxml==3.5.0
marrow.interface==1.0.1
marrow.mailer==4.0.0
marrow.util==1.2.3
MechanicalSoup==0.4.0
requests==2.9.1
six==1.10.0
